# docs/protocol/http/openapi-v1.json
{
  "components": {
    "parameters": {
...
      "RequestId": {
        "in": "header",
        "name": "X-Request-ID",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        },
        "description": "Stable typed failure.",
        "headers": {
          "X-Request-ID": {
            "description": "Matches the request and JSON body request_id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        }
      }
    },
    "schemas": {
...
      "JsonValue": {},
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Expose one bounded machine-readable failure without raw internals.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "error_code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "PtyOpenRequest": {
        "additionalProperties": false,
        "description": "Bind a new PTY to one active SSH session and initial geometry.",
        "properties": {
          "cols": {
            "maximum": 500,
            "minimum": 20,
            "title": "Cols",
            "type": "integer"
          },
          "rows": {
            "maximum": 300,
            "minimum": 5,
            "title": "Rows",
            "type": "integer"
          },
          "ssh_session_id": {
            "format": "uuid",
            "title": "Ssh Session Id",
            "type": "string"
          }
        },
        "required": [
          "ssh_session_id",
          "cols",
          "rows"
        ],
        "title": "PtyOpenRequest",
        "type": "object"
      },
      "PtyResizeRequest": {
        "additionalProperties": false,
        "description": "Describe a bounded PTY geometry update.",
        "properties": {
          "cols": {
            "maximum": 500,
            "minimum": 20,
            "title": "Cols",
            "type": "integer"
          },
          "rows": {
            "maximum": 300,
            "minimum": 5,
            "title": "Rows",
            "type": "integer"
          }
        },
        "required": [
          "cols",
          "rows"
        ],
        "title": "PtyResizeRequest",
        "type": "object"
      },
      "PtySession": {
        "additionalProperties": false,
        "description": "对桌面端公开的交互式 SSH PTY 会话快照。",
        "properties": {
          "cols": {
            "maximum": 500.0,
            "minimum": 20.0,
            "title": "Cols",
            "type": "integer"
          },
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "pty_session_id": {
            "format": "uuid",
            "title": "Pty Session Id",
            "type": "string"
          },
          "rows": {
            "maximum": 300.0,
            "minimum": 5.0,
            "title": "Rows",
            "type": "integer"
          },
          "ssh_session_id": {
            "format": "uuid",
            "title": "Ssh Session Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "OPEN",
              "CLOSED",
              "FAILED"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "pty_session_id",
          "ssh_session_id",
          "connection_id",
          "cols",
          "rows",
          "state"
        ],
        "title": "PtySession",
        "type": "object"
      },
      "PtySessionResponse": {
        "additionalProperties": false,
        "description": "Return one interactive PTY session snapshot.",
        "properties": {
          "pty_session": {
            "$ref": "#/components/schemas/PtySession"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "pty_session"
        ],
        "title": "PtySessionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Loopback-only API owned by the Rust-managed packaged child. The Runtime WebSocket is specified separately.",
    "title": "Harness Shell Private Python Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/pty/sessions": {
      "post": {
        "description": "Open one PTY channel on an explicitly selected SSH session.",
        "operationId": "open_pty_session_v1_pty_sessions_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtyOpenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PtySessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Open Pty Session"
      }
    },
    "/v1/pty/sessions/{pty_session_id}": {
      "delete": {
        "description": "Close one active PTY and return its terminal snapshot.",
        "operationId": "close_pty_session_v1_pty_sessions__pty_session_id__delete",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "pty_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pty Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PtySessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Close Pty Session"
      }
    },
    "/v1/pty/sessions/{pty_session_id}/resize": {
      "post": {
        "description": "Resize one active PTY without exposing its byte stream over HTTP.",
        "operationId": "resize_pty_session_v1_pty_sessions__pty_session_id__resize_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "in": "path",
            "name": "pty_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pty Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PtyResizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PtySessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Matches the request and JSON body request_id.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Resize Pty Session"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "8765"
        }
      }
    }
  ]
}